requirements.txt
streamlit
requests
beautifulsoup4
python-dotenv
fpdf
google-api-python-client